+++ src/index.ts
@@ -2,6 +2,13 @@ export * as EmbedAssertions from './messages/embed/Assertions';
 export * from './messages/embed/Embed';
 export * from './messages/formatters';
 
+export * as ComponentAssertions from './components/Assertions';
+export * from './components/button/BaseButton';
+export * from './components/button/InteractionButton';
+export * from './components/button/LinkButton';
+export * from './components/selectMenu/SelectMenu';
+export * from './components/selectMenu/SelectMenuOption';
+
 export * as SlashCommandAssertions from './interactions/slashCommands/Assertions';
 export * from './interactions/slashCommands/SlashCommandBuilder';
 export * from './interactions/slashCommands/SlashCommandSubcommands';
